requirements.txt
pymongo
fastapi
python-dotenv
pydantic
uvicorn[standard]